// src/types/project.ts
export interface Task {
  id: string;
  actionId?: number;
  assignedId?: number;
  assignedTo?: string;
  action: string;
  otherName?: string;
  dueDate: string;
  emailAlert: boolean;
  comment: string;
  isNew?: boolean;
}

export interface ProjectWizardData {
  projectId?: number;
  projectName: string;
  countryId: number;
  country?: string;
  state?: string;
  stateId: number;
  projectTypeId: number;
  roleId: number;
  customerTypeId: number;

  startDate: string;
  endDate: string;
  furnishingDates: string[];
  completionDate?: string;
  firstFurnishingDate?: string;
  lastFurnishingDate?: string;

  jobName: string;
  jobAddress: string;
  jobCity: string;
  jobStateId: number;
  jobZip: string;
  jobCountyId: number;
  jobCountryId: number;

  contractAmount: string;
  baseContractAmount: string;
  additionalCosts: string;
  revisedCost: string;
  paymentsCredits: string;
  unpaidBalance: string;
  jobProjectNumber: string;
  materialServicesDescription: string;

  selectedCustomerContacts: number;

  selectedProjectContacts: number[];

  documents: File[];
  uploaded_documents?: DocumentViewResponse[];

  tasks: Task[];
  removedTasks: string[];

  customerSignature: string;
  signatureDate: string;
}
// ...
export interface DocumentViewResponse {
  id: number;
  notes?: string;
  filename: string;
  date: string;
  title: string;
  project_id?: number;
  file_url: string;
  file_name: string;
  file_size: string;
  file_size_bytes: string;
  file_extension: string;
}